requirements.txt
supabase==2.24.0
python-dotenv==1.0.1
rich==13.7.1
tiktoken==0.7.0
websockets>=13.0
